Restructure files, split large python files
