+++ connector_magento/__init__.py
@@ -3,22 +3,6 @@
 # © 2016 Sodexis
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import setting
-from . import binding
-
 from . import components
-
-from . import magento_model
-from . import product
-from . import product_category
-from . import partner
-from . import partner_category
-from . import invoice
-from . import sale
-from . import delivery
-from . import stock_picking
-from . import stock_tracking
-from . import account_payment_mode
-
 from . import consumer
 from . import models
