+++ connector_magento/__manifest__.py
@@ -11,7 +11,6 @@
              'delivery',
              'sale_stock',
              'connector_ecommerce',
-             # 'product_multi_category',
              ],
  'external_dependencies': {
      'python': ['magento'],
@@ -27,18 +26,16 @@
             ],
  'data': ['data/connector_magento_data.xml',
           'security/ir.model.access.csv',
-          # 'views/setting_view.xml',
-          'views/magento_model_view.xml',
-          'views/product_view.xml',
-          'views/product_category_view.xml',
-          'views/partner_view.xml',
-          # 'views/sale_view.xml',
-          # 'views/invoice_view.xml',
+          'views/magento_backend_views.xml',
+          'views/product_views.xml',
+          'views/product_category_views.xml',
+          'views/partner_views.xml',
+          'views/invoice_views.xml',
           'views/connector_magento_menu.xml',
-          # 'views/delivery_view.xml',
-          # 'views/stock_view.xml',
-          # 'views/account_payment_mode_view.xml',
+          # 'views/delivery_views.xml',
+          'views/stock_views.xml',
+          'views/account_payment_mode_views.xml',
           ],
  'installable': True,
- 'application': True,
+ 'application': False,
  }
