+++ connector_magento/models/__init__.py
@@ -1,2 +1,17 @@
 # -*- coding: utf-8 -*-
+from . import magento_binding  # should be loaded first for inheritance
+
+from . import account_invoice
+from . import account_payment_mode
+from . import delivery
+from . import magento_backend
+from . import magento_store
+from . import magento_storeview
+from . import magento_website
+from . import partner
+from . import partner_category
+from . import product
+from . import product_category
 from . import queue_job
+from . import sale_order
+from . import stock_picking
