+++ connector_magento/models/account_invoice/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+from . import common
+from . import exporter
