+++ connector_magento/models/delivery/common.py
