+++ connector_magento/models/magento_backend/common.py
@@ -10,11 +10,9 @@ from contextlib import contextmanager
 from datetime import datetime, timedelta
 from odoo import models, fields, api, _
 from odoo.exceptions import UserError
-from odoo.addons.component.core import Component
 
-from odoo.addons.connector.components.mapper import mapping
 from odoo.addons.connector.checkpoint import checkpoint
-from .components.backend_adapter import  MagentoLocation, MagentoAPI
+from ...components.backend_adapter import MagentoLocation, MagentoAPI
 
 _logger = logging.getLogger(__name__)
 
@@ -401,315 +399,3 @@ class MagentoConfigSpecializer(models.AbstractModel):
             this.fiscal_position_id = (
                 this.specific_fiscal_position_id or
                 this._parent.fiscal_position_id)
-
-
-class MagentoWebsite(models.Model):
-    _name = 'magento.website'
-    _inherit = ['magento.binding', 'magento.config.specializer']
-    _description = 'Magento Website'
-    _parent_name = 'backend_id'
-
-    _order = 'sort_order ASC, id ASC'
-
-    name = fields.Char(required=True, readonly=True)
-    code = fields.Char(readonly=True)
-    sort_order = fields.Integer(string='Sort Order', readonly=True)
-    store_ids = fields.One2many(
-        comodel_name='magento.store',
-        inverse_name='website_id',
-        string='Stores',
-        readonly=True,
-    )
-    import_partners_from_date = fields.Datetime(
-        string='Import partners from date',
-    )
-    product_binding_ids = fields.Many2many(
-        comodel_name='magento.product.product',
-        string='Magento Products',
-        readonly=True,
-    )
-
-    @api.multi
-    def import_partners(self):
-        import_start_time = datetime.now()
-        for website in self:
-            backend = website.backend_id
-            if website.import_partners_from_date:
-                from_string = fields.Datetime.from_string
-                from_date = from_string(website.import_partners_from_date)
-            else:
-                from_date = None
-            self.env['magento.res.partner'].with_delay().import_batch(
-                backend,
-                filters={'magento_website_id': website.external_id,
-                         'from_date': from_date,
-                         'to_date': import_start_time}
-            )
-        # Records from Magento are imported based on their `created_at`
-        # date.  This date is set on Magento at the beginning of a
-        # transaction, so if the import is run between the beginning and
-        # the end of a transaction, the import of a record may be
-        # missed.  That's why we add a small buffer back in time where
-        # the eventually missed records will be retrieved.  This also
-        # means that we'll have jobs that import twice the same records,
-        # but this is not a big deal because they will be skipped when
-        # the last `sync_date` is the same.
-        next_time = import_start_time - timedelta(seconds=IMPORT_DELTA_BUFFER)
-        next_time = fields.Datetime.to_string(next_time)
-        self.write({'import_partners_from_date': next_time})
-        return True
-
-
-class MagentoStore(models.Model):
-    _name = 'magento.store'
-    _inherit = ['magento.binding', 'magento.config.specializer']
-    _description = 'Magento Store'
-    _parent_name = 'website_id'
-
-    name = fields.Char()
-    website_id = fields.Many2one(
-        comodel_name='magento.website',
-        string='Magento Website',
-        required=True,
-        readonly=True,
-        ondelete='cascade',
-    )
-    backend_id = fields.Many2one(
-        comodel_name='magento.backend',
-        related='website_id.backend_id',
-        string='Magento Backend',
-        store=True,
-        readonly=True,
-        # override 'magento.binding', can't be INSERTed if True:
-        required=False,
-    )
-    storeview_ids = fields.One2many(
-        comodel_name='magento.storeview',
-        inverse_name='store_id',
-        string="Storeviews",
-        readonly=True,
-    )
-    send_picking_done_mail = fields.Boolean(
-        string='Send email notification on picking done',
-        help="Does the picking export/creation should send "
-             "an email notification on Magento side?",
-    )
-    send_invoice_paid_mail = fields.Boolean(
-        string='Send email notification on invoice validated/paid',
-        help="Does the invoice export/creation should send "
-             "an email notification on Magento side?",
-    )
-    create_invoice_on = fields.Selection(
-        selection=[('open', 'Validate'),
-                   ('paid', 'Paid')],
-        string='Create invoice on action',
-        default='paid',
-        required=True,
-        help="Should the invoice be created in Magento "
-             "when it is validated or when it is paid in Odoo?\n"
-             "This only takes effect if the sales order's related "
-             "payment method is not giving an option for this by "
-             "itself. (See Payment Methods)",
-    )
-
-
-class MagentoStoreview(models.Model):
-    _name = 'magento.storeview'
-    _inherit = ['magento.binding', 'magento.config.specializer']
-    _description = "Magento Storeview"
-    _parent_name = 'store_id'
-
-    _order = 'sort_order ASC, id ASC'
-
-    name = fields.Char(required=True, readonly=True)
-    code = fields.Char(readonly=True)
-    enabled = fields.Boolean(string='Enabled', readonly=True)
-    sort_order = fields.Integer(string='Sort Order', readonly=True)
-    store_id = fields.Many2one(comodel_name='magento.store',
-                               string='Store',
-                               ondelete='cascade',
-                               readonly=True)
-    lang_id = fields.Many2one(comodel_name='res.lang', string='Language')
-    team_id = fields.Many2one(comodel_name='crm.team', string='Sales Team')
-    backend_id = fields.Many2one(
-        comodel_name='magento.backend',
-        related='store_id.website_id.backend_id',
-        string='Magento Backend',
-        store=True,
-        readonly=True,
-        # override 'magento.binding', can't be INSERTed if True:
-        required=False,
-    )
-    import_orders_from_date = fields.Datetime(
-        string='Import sale orders from date',
-        help='do not consider non-imported sale orders before this date. '
-             'Leave empty to import all sale orders',
-    )
-    no_sales_order_sync = fields.Boolean(
-        string='No Sales Order Synchronization',
-        help='Check if the storeview is active in Magento '
-             'but its sales orders should not be imported.',
-    )
-    catalog_price_tax_included = fields.Boolean(string='Prices include tax')
-
-    @api.multi
-    def import_sale_orders(self):
-        import_start_time = datetime.now()
-        for storeview in self:
-            if storeview.no_sales_order_sync:
-                _logger.debug("The storeview '%s' is active in Magento "
-                              "but is configured not to import the "
-                              "sales orders", storeview.name)
-                continue
-            backend = storeview.backend_id
-            if storeview.import_orders_from_date:
-                from_string = fields.Datetime.from_string
-                from_date = from_string(storeview.import_orders_from_date)
-            else:
-                from_date = None
-            delayable = self.env['magento.sale.order'].with_delay(priority=1)
-            filters = {
-                'magento_storeview_id': storeview.external_id,
-                'from_date': from_date,
-                'to_date': import_start_time,
-            }
-            delayable.import_batch(backend, filters=filters)
-        # Records from Magento are imported based on their `created_at`
-        # date.  This date is set on Magento at the beginning of a
-        # transaction, so if the import is run between the beginning and
-        # the end of a transaction, the import of a record may be
-        # missed.  That's why we add a small buffer back in time where
-        # the eventually missed records will be retrieved.  This also
-        # means that we'll have jobs that import twice the same records,
-        # but this is not a big deal because the sales orders will be
-        # imported the first time and the jobs will be skipped on the
-        # subsequent imports
-        next_time = import_start_time - timedelta(seconds=IMPORT_DELTA_BUFFER)
-        next_time = fields.Datetime.to_string(next_time)
-        self.write({'import_orders_from_date': next_time})
-        return True
-
-
-class WebsiteAdapter(Component):
-    _name = 'magento.website.adapter'
-    _inherit = 'magento.adapter'
-    _apply_on = 'magento.website'
-
-    _magento_model = 'ol_websites'
-    _admin_path = 'system_store/editWebsite/website_id/{id}'
-
-
-class StoreAdapter(Component):
-    _name = 'magento.store.adapter'
-    _inherit = 'magento.adapter'
-    _apply_on = 'magento.store'
-
-    _magento_model = 'ol_groups'
-    _admin_path = 'system_store/editGroup/group_id/{id}'
-
-
-class StoreviewAdapter(Component):
-    _name = 'magento.storeview.adapter'
-    _inherit = 'magento.adapter'
-    _apply_on = 'magento.storeview'
-
-    _magento_model = 'ol_storeviews'
-    _admin_path = 'system_store/editStore/store_id/{id}'
-
-
-class MetadataBatchImporter(Component):
-    """ Import the records directly, without delaying the jobs.
-
-    Import the Magento Websites, Stores, Storeviews
-
-    They are imported directly because this is a rare and fast operation,
-    and we don't really bother if it blocks the UI during this time.
-    (that's also a mean to rapidly check the connectivity with Magento).
-    """
-
-    _name = 'magento.metadata.batch.importer'
-    _inherit = 'magento.direct.batch.importer'
-    _apply_on = [
-        'magento.website',
-        'magento.store',
-        'magento.storeview',
-    ]
-
-
-class WebsiteImportMapper(Component):
-    _name = 'magento.website.mapper'
-    _inherit = 'magento.import.mapper'
-    _apply_on = 'magento.website'
-
-    direct = [('code', 'code'),
-              ('sort_order', 'sort_order')]
-
-    @mapping
-    def name(self, record):
-        name = record['name']
-        if name is None:
-            name = _('Undefined')
-        return {'name': name}
-
-    @mapping
-    def backend_id(self, record):
-        return {'backend_id': self.backend_record.id}
-
-
-class StoreImportMapper(Component):
-    _name = 'magento.store.mapper'
-    _inherit = 'magento.import.mapper'
-    _apply_on = 'magento.store'
-
-    direct = [('name', 'name')]
-
-    @mapping
-    def website_id(self, record):
-        binder = self.binder_for(model='magento.website')
-        binding = binder.to_internal(record['website_id'])
-        return {'website_id': binding.id}
-
-
-class StoreviewImportMapper(Component):
-    _name = 'magento.storeview.mapper'
-    _inherit = 'magento.import.mapper'
-    _apply_on = 'magento.storeview'
-
-    direct = [
-        ('name', 'name'),
-        ('code', 'code'),
-        ('is_active', 'enabled'),
-        ('sort_order', 'sort_order'),
-    ]
-
-    @mapping
-    def store_id(self, record):
-        binder = self.binder_for(model='magento.store')
-        binding = binder.to_internal(record['group_id'])
-        return {'store_id': binding.id}
-
-
-class StoreImporter(Component):
-    """ Import one Magento Store (create a sale.shop via _inherits) """
-
-    _name = 'magento.store.importer'
-    _inherit = 'magento.importer'
-    _apply_on = 'magento.store'
-
-    def _create(self, data):
-        binding = super(StoreImporter, self)._create(data)
-        self.backend_record.add_checkpoint(binding)
-        return binding
-
-
-class StoreviewImporter(Component):
-    """ Import one Magento Storeview """
-
-    _name = 'magento.storeview.importer'
-    _inherit = 'magento.importer'
-    _apply_on = 'magento.storeview'
-
-    def _create(self, data):
-        binding = super(StoreviewImporter, self)._create(data)
-        self.backend_record.add_checkpoint(binding)
-        return binding
