+++ connector_magento/models/magento_binding/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from . import common
