+++ connector_magento/models/partner/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+from . import common
+from . import importer
