+++ connector_magento/models/product/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+from . import common
+from . import importer
