+++ connector_magento/models/queue_job/common.py
