+++ connector_magento/models/sale_order/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+from . import common
+from . import importer
+from . import exporter
