+++ connector_magento/models/stock_picking/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+from . import common
+from . import exporter
+from . import tracking_exporter
