+++ connector_magento/views/account_payment_mode_views.xml
