+++ connector_magento/views/delivery_views.xml
