+++ connector_magento/views/invoice_views.xml
