+++ connector_magento/views/magento_backend_views.xml
