+++ connector_magento/views/partner_views.xml
