+++ connector_magento/views/product_category_views.xml
