+++ connector_magento/views/product_views.xml
