--- connector_magento/views/sale_view.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-
-</odoo>
