+++ connector_magento/views/stock_views.xml
